restructured diagram generation
